# openapi.yaml
openapi: 3.1.0
info:
  title: Client Onboarding MCP REST API
  version: 1.0.0
  description: REST facade for client onboarding, basic info, facility limit, and outreach tools exposed by the MCP server.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://example.ngrok-free.app
    description: Replace with your HTTPS tunnel or hosted endpoint
tags:
  - name: client-tools
    description: Client onboarding tools
paths:
  /health:
    get:
      operationId: get_health
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  service:
                    type: string
                  version:
                    type: string
                  api_version:
                    type: string
                  backend:
                    type: string
  /api/v1/clients/{client_id}/onboarding-status:
    get:
      operationId: get_client_onboarding_status
      tags: [client-tools]
      summary: Get client onboarding status
      description: Returns the hardcoded onboarding status for a client ID.
      x-oai-isConsequential: false
      x-openai-isConsequential: false
      parameters:
        - name: client_id
          in: path
          required: true
          description: Client identifier, for example 123.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Client onboarding status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingStatusResponse"
              examples:
                completed:
                  value:
                    client_id: "123"
                    status: Completed
                    message: Client onboarding completed successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
  /api/v1/clients/{client_id}/basic-info/{field}:
    get:
      operationId: get_client_basic_info
      tags: [client-tools]
      summary: Get client basic info
      description: Returns the hardcoded name or family value for a client ID.
      x-oai-isConsequential: false
      x-openai-isConsequential: false
      parameters:
        - name: client_id
          in: path
          required: true
          description: Client identifier, for example 123.
          schema:
            type: string
            minLength: 1
        - name: field
          in: path
          required: true
          description: Allowed fields are name and family.
          schema:
            type: string
            enum: [name, family]
      responses:
        "200":
          description: Client basic info response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BasicInfoResponse"
              examples:
                name:
                  value:
                    client_id: "123"
                    field: name
                    value: John
                family:
                  value:
                    client_id: "123"
                    field: family
                    value: Doe
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
  /api/v1/clients/{client_id}/facility:
    get:
      operationId: get_client_facility_limit
      tags: [client-tools]
      summary: Get client facility limit
      description: Returns the client name and current facility limit in EUR for a client ID.
      x-oai-isConsequential: false
      x-openai-isConsequential: false
      parameters:
        - name: client_id
          in: path
          required: true
          description: Client identifier, for example 123.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Client facility limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientFacilityResponse"
              example:
                client_id: "123"
                client_name: John Doe
                facility_limit_eur: 75000000
                currency: EUR
                formatted_limit: EUR 75,000,000
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
  /api/v1/clients/{client_id}/outreach-summary:
    get:
      operationId: summarize_client_outreach
      tags: [client-tools]
      summary: Summarize client outreach
      description: Summarizes outreach reasons, highlights, and the total number of questions that should be answered for a client ID.
      x-oai-isConsequential: false
      x-openai-isConsequential: false
      parameters:
        - name: client_id
          in: path
          required: true
          description: Client identifier, for example 123.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Client outreach summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/BackendUnavailable"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Placeholder API key authentication. Enable with MCP_AUTH_REQUIRED=true.
  schemas:
    OnboardingStatusResponse:
      type: object
      required: [client_id, status, message]
      properties:
        client_id:
          type: string
          examples: ["123"]
        status:
          type: string
          examples: [Completed]
        message:
          type: string
          examples: [Client onboarding completed successfully]
    BasicInfoResponse:
      type: object
      required: [client_id, field, value]
      properties:
        client_id:
          type: string
          examples: ["123"]
        field:
          type: string
          enum: [name, family]
        value:
          type: string
          examples: [John]
    ClientFacilityResponse:
      type: object
      required: [client_id, client_name, facility_limit_eur, currency, formatted_limit]
      properties:
        client_id:
          type: string
          examples: ["123"]
        client_name:
          type: string
          examples: [John Doe]
        facility_limit_eur:
          type: integer
          minimum: 1000000
          maximum: 200000000
          examples: [75000000]
        currency:
          type: string
          examples: [EUR]
        formatted_limit:
          type: string
          examples: ["EUR 75,000,000"]
    OutreachItem:
      type: object
      required: [outreach_id, date, channel, reason, questions_to_answer]
      properties:
        outreach_id:
          type: string
        date:
          type: string
        channel:
          type: string
        reason:
          type: string
        questions_to_answer:
          type: array
          items:
            type: string
    OutreachSummaryResponse:
      type: object
      required:
        - client_id
        - client_name
        - outreach_count
        - questions_to_answer_count
        - reasons
        - highlights
        - summary
        - outreaches
      properties:
        client_id:
          type: string
          examples: ["123"]
        client_name:
          type: string
          examples: [John Doe]
        outreach_count:
          type: integer
          examples: [2]
        questions_to_answer_count:
          type: integer
          examples: [5]
        reasons:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        summary:
          type: string
        outreaches:
          type: array
          items:
            $ref: "#/components/schemas/OutreachItem"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - array
                - string
                - "null"
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: Missing or invalid x-api-key header.
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: validation_error
              message: Request validation failed.
              details: []
    BackendUnavailable:
      description: Mock backend is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: backend_unavailable
              message: Mock backend server is not reachable.